python-10/requirements.txt
appdirs==1.4.4
astroid==2.4.1     
atomicwrites==1.4.0
attrs==19.3.0      
backcall==0.1.0    
certifi==2018.10.15
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
distlib==0.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.2
django-bootstrap-form==3.4
django-password-validation==0.1.1
djangorestframework==3.11.0
filelock==3.0.12
flake8==3.8.1
idna==2.9
importlib-metadata==1.6.0
ipython==7.15.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
jwcrypto==0.7
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.3.0
packaging==20.4
parso==0.7.0
pickleshare==0.7.5
Pillow==5.3.0
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2==2.7.5
py==1.8.1
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.2
pyparsing==2.4.7
pytest==5.4.2
python-decouple==3.1
python-jwt==3.2.6
pytz==2018.5
requests==2.23.0
selenium==3.141.0
six==1.14.0
static3==0.7.0
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.20
wcwidth==0.1.9
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0